% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{NetGraph}
\alias{NetGraph}
\title{Converts the whale network from a network object to a graph object}
\usage{
NetGraph(net)
}
\arguments{
\item{net}{A network object, containing the nodal attributes used by Load2015Net()}
}
\value{
A graph object version of the network
}
\description{
Converts the whale network from a network object to a graph object
}

