% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{simresults.vioplot}
\alias{simresults.vioplot}
\title{Produce a violin plot of the results of simulate.nonkin}
\usage{
simresults.vioplot(simresults)
}
\arguments{
\item{simresults}{A matrix produced by simulate.nonkin}
}
\value{
A violin plot of the results, with observed values plotted in red
}
\description{
Produce a violin plot of the results of simulate.nonkin
}

