% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{simulate.nonkin}
\alias{simulate.nonkin}
\title{Test for unusually high number of cross-pod dyad types based on an ergm via simulation}
\usage{
\method{simulate}{nonkin}(model, reps = 10000)
}
\arguments{
\item{model}{The ergm to simulate from}

\item{reps}{The number of networks to simulate. Defaults to 10,000.}
}
\value{
A matrix containing the number of total cross-pod dyads, and dyads of each type, for each simulation
}
\description{
Test for unusually high number of cross-pod dyad types based on an ergm via simulation
}

