% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{whale.community.gof}
\alias{whale.community.gof}
\title{Goodness of fit diagnostics for ERGM based on community structure}
\usage{
whale.community.gof(model, sample.size = 500)
}
\arguments{
\item{model}{An ergm object}

\item{sample.size}{The number of simulations to perform. Defaults to 500}
}
\value{
A plot of densities of number of components, communities, and modularity based on the ergm, compared to initial network.
}
\description{
Goodness of fit diagnostics for ERGM based on community structure
}

