% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{WhaleNetGraph}
\alias{WhaleNetGraph}
\title{Converts the whale network from a network object to a graph object}
\usage{
WhaleNetGraph(net)
}
\arguments{
\item{net}{A network object, containing the nodal attributes used by Load2015Net()}
}
\value{
A graph object version of the network
}
\description{
Converts the whale network from a network object to a graph object
}
\details{
This function converts networks to a graph objects to allow for community detection using the igraph package.
}

