% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{Load2015Net}
\alias{Load2015Net}
\title{Load the 2015 SRKW Preferred Association Network}
\usage{
Load2015Net()
}
\value{
A network object of the 2015 preferred associations with nodal attributes
}
\description{
Load the 2015 SRKW Preferred Association Network
}

