% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{simresults.vioplot}
\alias{simresults.vioplot}
\title{Produce a violin plot of the results of simulate.nonkin}
\usage{
simresults.vioplot(simresults, point.col = "red", vio.col = "grey")
}
\arguments{
\item{simresults}{A matrix produced by simulate.nonkin}

\item{point.col}{The color of the points used to indicate observed values. Defaults to red.}

\item{vio.col}{The color of the violin plots indicating simulation results. Defaults to grey.}
}
\value{
A violin plot of the results, with observed values plotted in red
}
\description{
Produce a violin plot of the results of simulate.nonkin
}

