% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srkw.sna.code.R
\name{whale.ergm}
\alias{whale.ergm}
\title{Create the ergm used in the original analysis}
\usage{
whale.ergm(net)
}
\arguments{
\item{net}{A network object}
}
\value{
An ergm object
}
\description{
Create the ergm used in the original analysis
}

